+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- Add prop 'isOpenOnMount' to allow opening submenu on mount
+- Add prop 'onMountBehavior' to allow opening submenu on mount
 
 ## [2.32.0] - 2021-06-24
 ### Added